requirements.txt
kafka-python==1.4.4
python-snappy==0.5.4
